requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
python-dotenv==1.0.1
pydantic==2.10.4
pytest==8.3.4
pytest-cov==6.0.0
httpx==0.28.1
requests==2.32.3
